Add the crawl method and use it generically through the abstract PriceIndexReader class instead of the specialized classes
