+++ implementation/consultasbr/src/main/java/br/com/quintoandar/consultasbr/priceindex/IGPMReader.java
@@ -1,0 +1,25 @@
+package br.com.quintoandar.consultasbr.priceindex;
+
+/**
+ * @author <a href="mpereira@quintoandar.com.br">Moacyr</a>
+ **/
+public class IGPMReader extends PriceIndexReader {
+
+	@Override
+	protected String getJavascriptFilename() {
+		return "igpmf";
+	}
+
+	@Override
+	protected int getNumMesesOffset() {
+		return 1;
+	}
+
+	public static void main(String[] args) {
+		IGPMReader r = new IGPMReader();
+		r.crawl();
+		System.out.println(r.getMes());
+		System.out.println(r.getAcumulado());
+		System.out.println(r.getMensal());
+	}
+}
